move part of APIs from utils/ValidationUtil to config/Utils
