+++ core/src/config/Utils.h
@@ -12,12 +12,49 @@
 #pragma once
 
 #include <string>
+#include <vector>
+
+#include "db/Types.h"
+#include "db/meta/MetaTypes.h"
+#include "utils/Json.h"
+#include "utils/Status.h"
 
 namespace milvus {
 namespace server {
 
-int64_t
+extern int64_t
 parse_bytes(const std::string& str, std::string& err);
 
+extern Status
+ValidateGpuIndex(int32_t gpu_index);
+
+#ifdef MILVUS_GPU_VERSION
+extern Status
+GetGpuMemory(int32_t gpu_index, int64_t& memory);
+#endif
+
+extern Status
+ValidateIpAddress(const std::string& ip_address);
+
+extern Status
+ValidateStringIsNumber(const std::string& str);
+
+extern Status
+ValidateStringIsBool(const std::string& str);
+
+extern Status
+ValidateStringIsFloat(const std::string& str);
+
+extern Status
+ValidateDbURI(const std::string& uri);
+
+extern Status
+ValidateStoragePath(const std::string& path);
+
+extern Status
+ValidateLogLevel(const std::string& level);
+
+extern bool
+IsNumber(const std::string& s);
 }  // namespace server
 }  // namespace milvus
