+++ core/src/db/meta/MySQLMetaImpl.cpp
@@ -38,7 +38,6 @@
 #include "utils/Exception.h"
 #include "utils/Log.h"
 #include "utils/StringHelpFunctions.h"
-#include "utils/ValidationUtil.h"
 
 namespace milvus {
 namespace engine {
