+++ core/src/scheduler/task/BuildIndexTask.cpp
@@ -26,7 +26,6 @@
 #include "utils/Exception.h"
 #include "utils/Log.h"
 #include "utils/TimeRecorder.h"
-#include "utils/ValidationUtil.h"
 
 namespace milvus {
 namespace scheduler {
