+++ core/src/utils/StringHelpFunctions.cpp
@@ -16,8 +16,6 @@
 #include <regex>
 #include <string>
 
-#include "utils/ValidationUtil.h"
-
 namespace milvus {
 namespace server {
 
@@ -156,11 +154,6 @@ StringHelpFunctions::IsRegexMatch(const std::string& target_str, const std::stri
 
 Status
 StringHelpFunctions::ConvertToBoolean(const std::string& str, bool& value) {
-    auto status = ValidationUtil::ValidateStringIsBool(str);
-    if (!status.ok()) {
-        return status;
-    }
-
     std::string s = str;
     std::transform(s.begin(), s.end(), s.begin(), ::tolower);
     value = s == "true" || s == "on" || s == "yes" || s == "1";
