+++ core/unittest/server/test_config.cpp
@@ -22,7 +22,6 @@
 #include "server/utils.h"
 #include "utils/CommonUtil.h"
 #include "utils/StringHelpFunctions.h"
-#include "utils/ValidationUtil.h"
 
 namespace {
